# spec/openapi.yaml
openapi: 3.0.3
info:
  title: Seorilabs Platform API
  version: 1.0.0-draft
  description: |
    Seorilabs 공통 플랫폼 API.

    ## 계약 규칙

    - **이 문서가 계약의 source of truth다.** Go 타입도 TS 타입도 여기서 생성된 산출물이다
    - **`/v1`은 영구히 깨지지 않는다.** 마켓 배포된 구버전 SDK가 2~3년 산다
    - 필드 **추가만** 허용한다. 제거와 의미 변경은 `/v2`
    - 양방향 forward-compat — 서버는 요청의 미지 필드를 무시하고, SDK는 응답의 미지 필드를 무시한다
    - 모든 응답은 envelope로 감싼다. 성공은 `{ok:true, result}`, 실패는 `{ok:false, error}`
    - 요청 본문 상한 64KiB. 제어문자 거부. 허용 필드 외 어떤 키도 400

    상태: **D0 초안.** 스키마 세부는 각 단계에서 확정한다.
servers:
  - url: https://platform-api.seorilabs.example/v1
    description: 확정 필요 — P0 배포 후 실제 URL로 교체
...
paths:
  /auth/session:
    post:
      tags: [auth]
      summary: 세션 발급
      description: |
        앱별 자격증명을 플랫폼 세션 토큰으로 교환한다.

        부팅 시 왕복을 1회로 줄이기 위해 RemoteConfig와 SDK 상태를 함께 반환한다.
        Godot의 HTTPRequest는 동시 1요청만 처리하므로 이게 중요하다.
      operationId: createSession
      security: []
      parameters:
        - $ref: '#/components/parameters/SeoriApp'
        - $ref: '#/components/parameters/AppCheck'
        - $ref: '#/components/parameters/SeoriSdk'
        - $ref: '#/components/parameters/SeoriRuntime'
        - $ref: '#/components/parameters/SeoriAppVer'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSessionRequest'
      responses:
        '200':
          description: 발급 성공
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/EnvelopeOk'
                  - type: object
                    properties:
                      result:
                        $ref: '#/components/schemas/Session'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /auth/refresh:
    post:
      tags: [auth]
      summary: 세션 갱신
      operationId: refreshSession
      security: []
      parameters:
        - $ref: '#/components/parameters/SeoriApp'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken:
                  type: string
                  maxLength: 512
      responses:
        '200':
          description: 갱신 성공
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/EnvelopeOk'
                  - type: object
                    properties:
                      result:
                        $ref: '#/components/schemas/Session'
        '401': { $ref: '#/components/responses/Unauthorized' }
  /auth/firebase-custom-token:
    post:
      tags: [auth]
      summary: Firebase custom token bridge
      description: |
        platform이 검증하거나 새로 만든 앱 사용자 uid로 Firebase custom token을 발급한다.

        기존 Firebase ID token을 보내면 그 uid를 그대로 사용해 기존 Firestore 소유권을
        보존한다. 토큰이 없으면 uid는 서버가 생성한다. 클라이언트가 uid를 직접 고를 수
        없으며, 이 응답의 custom token은 즉시 `signInWithCustomToken`에 사용하고 저장하지 않는다.
      operationId: createFirebaseCustomToken
      security: []
      parameters:
        - $ref: '#/components/parameters/SeoriApp'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateFirebaseCustomTokenRequest'
      responses:
        '200':
          description: 발급 성공
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/EnvelopeOk'
                  - type: object
                    properties:
                      result:
                        $ref: '#/components/schemas/FirebaseCustomToken'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /auth/firebase-account:
    delete:
      tags: [auth]
      summary: Firebase 사용자 Platform 매핑 삭제
      description: |
        앱의 Firebase ID token을 검증하고 해당 uid의 Platform identity 매핑과
        PII 없는 사용자 문서를 삭제한다. Firebase Auth와 앱 데이터 삭제는 앱
        프로젝트가 이어서 수행한다. IAP 감사 원장은 삭제하지 않는다.
      operationId: deleteFirebaseAccount
      security: []
      parameters:
        - $ref: '#/components/parameters/SeoriApp'
        - $ref: '#/components/parameters/AppCheck'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [appId, firebaseIdToken]
              additionalProperties: false
              properties:
                appId: { type: string, pattern: '^[a-z0-9][a-z0-9-]{0,63}$' }
                firebaseIdToken:
                  type: string
                  minLength: 1
                  maxLength: 4096
      responses:
        '200':
          description: 삭제 성공
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/EnvelopeOk'
                  - type: object
                    properties:
                      result:
                        type: object
                        required: [deleted]
                        properties:
                          deleted: { type: boolean }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
...
  /users/me:
    delete:
      tags: [user]
      summary: 플랫폼 사용자 삭제
      description: |
        앱이 계정을 삭제할 때 호출한다.
        PII를 저장하지 않더라도 삭제 경로는 있어야 한다.

        **IAP 원장 문서는 삭제하지 않는다.** 소유자 참조만 끊는다.
      operationId: deleteCurrentUser
      responses:
        '200':
          description: 삭제됨
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/EnvelopeOk'
                  - type: object
                    properties:
                      result:
                        type: object
                        properties:
                          deleted: { type: boolean }
        '401': { $ref: '#/components/responses/Unauthorized' }

  # 마켓별 엔드포인트를 두지 않는다. 증빙 필드 이름도 마켓마다 다르지
  # 않고 token 하나다. 마켓이 늘 때마다 경로와 필드가 늘면 SDK가 마켓
  # 개수만큼 갈라지고, 그 갈라짐은 지급 여부가 걸린 자리에서 드러난다.
...
components:
  securitySchemes:
    PlatformToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: '`/auth/session`이 발급한 플랫폼 세션 토큰'
    PubSubOidc:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Pub/Sub push subscription이 붙이는 Google OIDC 토큰
    AdminOidc:
      type: http
      scheme: bearer
      bearerFormat: Google OIDC ID token
      description: |
        Cloud Run Admin audience를 대상으로 발급한 Google OIDC ID token.
        GET은 read 또는 write allowlist, mutation은 write allowlist만 허용한다.
  parameters:
    SeoriApp:
      name: X-Seori-App
      in: header
      required: true
      description: |
        앱 식별자. **권한이 아니라 어느 프로젝트로 검증할지 고르는 힌트다.**
        헤더를 바꿔도 토큰의 aud 불일치로 거부된다.
      schema:
        type: string
        pattern: '^[a-z0-9][a-z0-9-]{0,63}$'
    SeoriSdk:
      name: X-Seori-Sdk
      in: header
      description: 'SDK 종류와 버전. 예 `gd/0.3.1`, `ts/0.3.1`. 구버전 트래픽 관측에 쓴다'
      schema: { type: string, maxLength: 32 }
    SeoriRuntime:
      name: X-Seori-Runtime
      in: header
      description: '`godot-native-android`, `godot-web-ait`, `rn-ios`, `ait-rn`, `web`'
      schema: { type: string, maxLength: 32 }
    SeoriAppVer:
      name: X-Seori-AppVer
      in: header
      description: 앱 버전. RemoteConfig 타겟팅에 쓴다
      schema: { type: string, maxLength: 32 }
    AppCheck:
      name: X-Firebase-AppCheck
      in: header
      description: |
        레지스트리 `require_app_check`가 true인 앱에서만 필수다.

        1단계는 optional로 시작한다. Godot에 App Check SDK가 없고
        replay 방지(consume)가 Go SDK에 없다.
      schema: { type: string }
...
  schemas:
    EnvelopeOk:
      type: object
      required: [ok]
      properties:
        ok:
          type: boolean
          enum: [true]
    EnvelopeError:
      type: object
      required: [ok, error]
      properties:
        ok:
          type: boolean
          enum: [false]
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: snake_case 에러 코드. 전체 목록은 docs/03-architecture/iap.md
              example: purchase_owned_by_another_user
            message:
              type: string
              description: 한국어 메시지. 클라이언트가 이 문자열로 분기하지 않는다
...
    CreateSessionRequest:
      type: object
      required: [appId, credential]
      additionalProperties: false
      properties:
        appId:
          type: string
          pattern: '^[a-z0-9][a-z0-9-]{0,63}$'
        credential:
          type: object
          required: [kind, value]
          additionalProperties: false
          properties:
            kind:
              type: string
              enum: [firebase-id-token, ait-login, anonymous]
              description: |
                **anonymous는 사칭 가능한 신원이다.**
                RemoteConfig 조회와 이벤트 로그만 허용하고 IAP는 금지한다.
            value:
              type: string
              maxLength: 4096
            referrer:
              type: string
              enum: [DEFAULT, SANDBOX]
              description: ait-login에서만 필수인 appLogin referrer
...
    CreateFirebaseCustomTokenRequest:
      type: object
      required: [appId]
      additionalProperties: false
      properties:
        appId:
          type: string
          pattern: '^[a-z0-9][a-z0-9-]{0,63}$'
        existingFirebaseIdToken:
          type: string
          maxLength: 4096
          description: 기존 사용자의 uid 보존이 필요할 때만 보낸다
    FirebaseCustomToken:
      type: object
      required: [firebaseCustomToken, appUserId]
      properties:
        firebaseCustomToken:
          type: string
          description: Firebase signInWithCustomToken에 한 번 사용하고 저장하지 않는다
        appUserId:
          type: string
          minLength: 1
          maxLength: 128
          description: custom token의 Firebase uid. 기존 토큰을 보낸 경우 기존 uid와 같아야 한다
    Session:
      type: object
      required:
        [platformToken, refreshToken, platformUserId, supportCode, expiresIn, serverTimeUnix, features, sdk]
      properties:
        platformToken: { type: string }
        refreshToken: { type: string }
        platformUserId:
          type: string
          pattern: '^pu_[0-9A-HJKMNP-TV-Z]{26}$'
          description: 'pu_ + ULID. firebase_uid에서 파생하지 않는다'
        supportCode:
          type: string
          pattern: '^[A-Z]{1,3}-[0-9A-HJKMNP-TV-Z]{8}$'
          description: >
            유저가 문의에 첨부하는 식별자. 앱 설정 화면이 이 값을 보여준다.
            Firebase uid를 보여주면 CS가 그걸로 원장을 찾을 수 없다 —
            플랫폼은 앱의 uid를 조회 키로 두지 않고 PII도 저장하지 않아
            이메일 검색이 성립하지 않는다. Admin API가 이 코드로 조회한다.
        appUserId:
          type: string
          description: 앱 쪽 식별자. 디버깅용
        expiresIn:
          type: integer
          description: 초
        serverTimeUnix:
          type: integer
          format: int64
        isAnonymous:
          type: boolean
          description: true면 IAP 경로가 막힌다
        features:
          $ref: '#/components/schemas/Features'
        sdk:
          $ref: '#/components/schemas/SdkStatus'
        maintenance:
          $ref: '#/components/schemas/Maintenance'
        configEtag:
          type: string
          description: 이 값이 그대로면 /config를 호출할 필요가 없다
    Features:
      type: object
      additionalProperties:
        type: boolean
      example:
        iap: true
        events: true
        config: true
    SdkStatus:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [ok, deprecated, blocked]
          description: |
            **마켓 배포된 구버전 SDK를 서버에서 끄는 유일한 수단이다.**
            blocked면 SDK가 이벤트만 남기고 나머지를 정지한다.
        message: { type: string }
        updateUrl: { type: string }
        minSupportedVersion: { type: string }
    Maintenance:
      type: object
      required: [active]
      properties:
        active: { type: boolean }
        message: { type: string }
        until: { type: string, format: date-time }
...
  responses:
    BadRequest:
      description: 요청 형식 오류. 허용 필드 외 키가 있어도 여기로 온다
      content:
        application/json:
          schema: { $ref: '#/components/schemas/EnvelopeError' }
    Unauthorized:
      description: 인증 실패
      content:
        application/json:
          schema: { $ref: '#/components/schemas/EnvelopeError' }
...
    Forbidden:
      description: 권한 부족, 앱 paused, anonymous 또는 관리 allowlist 불일치
      content:
        application/json:
          schema: { $ref: '#/components/schemas/EnvelopeError' }
...
    RateLimited:
      description: 한도 초과
      content:
        application/json:
          schema: { $ref: '#/components/schemas/EnvelopeError' }
...
    ServiceUnavailable:
      description: 설정 오류이거나 일시적 처리 불가
      content:
        application/json:
          schema: { $ref: '#/components/schemas/EnvelopeError' }
...
security:
  - PlatformToken: []